requirements.txt
flask
requests
python-dotenv
PyPDF2
pytesseract
Pillow
gunicorn